requirements.txt
async-timeout==4.0.2
certifi==2022.6.15
charset-normalizer==2.1.0
Deprecated==1.2.13
idna==3.3
packaging==21.3
pyparsing==3.0.9
pyTelegramBotAPI==4.6.0
python-dateutil==2.8.2
redis==4.3.4
requests==2.28.1
six==1.16.0
urllib3==1.26.10
wrapt==1.14.1
